java/syncbase: Fix AccessList Tags
When applying deltas, we applied ALL the tags at once. However, we
must instead filter the tags we pass to collections and syncgroups to
not include unnecessary tags (to avoid an error).

We were also not very careful when parsing the Permissions map and
AccessList's IN/NOT_IN. I've tried to be more careful by initializing
wherever necessary.

Change-Id: Iddf23ab17d245bef500e9caa6136a88687bb0898