java/syncbase: Test Advertise+Scan. Refactor Android Helpers
I was changing a small thing about Advertise/Scan in Neighborhood.
However, this lead me to discover that the tests were not longer
compiling.

So I refactored the Android Helpers to the side (to avoid
importing things that the tests did not understand).

In addition to the new tests (and old ones) passing, dice roller was
updated to confirm that the Android Helper changes are okay.

Change-Id: I7822c9d9ea4b9ddaf99bedc694c3688904456c50